requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pipenv==2021.5.29
platformdirs==2.2.0
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
